Har fjernet DummyDropDown basert på tabell som ikke skal leve og BehandleInnmeldingSaksB består nå bare av skjelettet
