+++ Gruppe10KVprototype/Views/Saksbehandler/DropDownDummySaksB/DropDownDummySaksB.cshtml
@@ -1,0 +1,59 @@
+﻿@model ViewModels.DropDownDummySaksBViewModel
+
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Rapport</title>
+</head>
+<body>
+    <h1>Lene Olsen</h1>
+    <h1>- har drept skjelettet</h1>
+    <h1>nb! databasen har bare 9 oppføringer</h1>
+
+    @if (Model != null)
+    {
+        <form asp-action="DropDownDummySaksB" method="post">
+            <div>
+                <label for="Innmelding_InnmeldID">Innmeldings ID:</label>
+                <input type="number" asp-for="InnmeldingE.InnmeldID" required />
+                <button type="submit">Hent Innmelding</button>
+            </div>
+        </form>
+        @if (Model.InnmeldingE != null)
+        {
+            @if (Model.InnmeldingE.InnmeldID > 0)
+            {
+                <div>
+                    <h2>Innmelding Detaljer:</h2>
+                    <p>InnmeldID: @Model.InnmeldingE.InnmeldID</p>
+                    <p>InnmelderID: @Model.InnmeldingE.InnmelderID</p>
+                    <p>Innmeldingstidspunkt: @Model.InnmeldingE.Innmeldingstidspunkt</p>
+                    <p>Nåværende Status: @Model.InnmeldingE.StatusID</p>
+                    <p>SaksbehandlerID: @(Model.InnmeldingE.SaksbehandlerID ?? 0)</p>
+                    <p>Siste Endring: @Model.InnmeldingE.SisteEndring</p>
+                    <p>KommuneNR: @Model.InnmeldingE.KommuneNR</p>
+                    <p>Prioritet: @Model.InnmeldingE.PrioritetID</p>
+                    <p>KartID: @Model.InnmeldingE.KartID</p>
+
+                    <div>
+                        <form asp-action="UpdateStatus" method="post">
+                            <input type="hidden" asp-for="InnmeldingE.InnmeldID" />
+                            <label for="status">Velg status:</label>
+                            <select asp-for="InnmeldingE.StatusID" asp-items="Model.StatusList" class="form-control"></select>
+                            <button type="submit" class="btn btn-primary">Oppdater status</button>
+                        </form>
+                    </div>
+                </div>
+            }
+            else
+            {
+                <p>Ingen innmelding funnet med den angitte ID.</p>
+            }
+        }
+    }
+    else
+    {
+        <p>Modellen er null. Vennligst kontakt systemadministrator.</p>
+    }
+</body>
+</html>
