+++ Gruppe10KVprototype/Views/Shared/_Layout.cshtml
@@ -27,34 +27,35 @@
                             <a class="nav-link" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" asp-controller="OversiktAlleInnmeldingerSaksBIncidentForm" asp-action="OversiktAlleInnmeldingerSaksBIncidentForm">Saksbehandler</a>
+                            <a class="nav-link" asp-controller="OversiktAlleInnmeldingerSaksBIncidentForm" asp-action="OversiktAlleInnmeldingerSaksBIncidentForm">OversiktAlleInnmeldingerSaksBIncidentForm</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" asp-area="" asp-controller="MineInnmeldinger" asp-action="MineInnmeldinger">Mine Saker</a>
+                            <a class="nav-link" asp-area="" asp-controller="MineInnmeldinger" asp-action="MineInnmeldinger">MineInnmedinger</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" asp-area="" asp-controller="OppdatereInnmelding" asp-action="OppdatereInnmelding">Oppdater Innmelding</a>
+                            <a class="nav-link" asp-area="" asp-controller="OppdatereInnmelding" asp-action="OppdatereInnmelding">OppdatereInnmelding</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" asp-area="" asp-controller="LoggInn" asp-action="LoggInn">Logg Inn</a>
+                            <a class="nav-link" asp-area="" asp-controller="LoggInn" asp-action="LoggInn">LoggInn</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" asp-area="" asp-controller="LandingsSide" asp-action="LandingsSide">Landings Side</a>
+                            <a class="nav-link" asp-area="" asp-controller="LandingsSide" asp-action="LandingsSide">LandingsSide</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" asp-area="" asp-controller="RegistrerDeg" asp-action="RegistrerDeg">Registrer Deg</a>
+                            <a class="nav-link" asp-area="" asp-controller="RegistrerDeg" asp-action="RegistrerDeg">RegistrerDeg</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" asp-area="" asp-controller="KnyttInnmeldigTilPerson" asp-action="KnyttInnmeldigTilPerson">Knytt Innmelding Til Person</a>
+                            <a class="nav-link" asp-area="" asp-controller="KnyttInnmeldigTilPerson" asp-action="KnyttInnmeldigTilPerson">KnyttInnmeldingTilPerson</a>
                         </li>
                         <div >
-                            <a href="@Url.Action("LandingsSideSaksB", "LandingsSideSaksB")">LandingsSide SaksB</a>
-                            <a href="@Url.Action("BehandleInnmeldingSaksB", "BehandleInnmeldingSaksB")">BEHANDLE INNMELDING SaksB</a>
-                            <a href="@Url.Action("OversiktAlleInnmeldingerSaksB", "OversiktAlleInnmeldingerSaksB")">Alle Innmeldinger SaksB</a>
+                            <a href="@Url.Action("LandingsSideSaksB", "LandingsSideSaksB")">LandingsSideSaksB</a>
+                            <a href="@Url.Action("BehandleInnmeldingSaksB", "BehandleInnmeldingSaksB")">BehandleInnmeldingSaksB</a>
+                            <a href="@Url.Action("OversiktAlleInnmeldingerSaksB", "OversiktAlleInnmeldingerSaksB")">OversiktAlleInnmeldingerSaksB</a>
                             <a href="@Url.Action("LogginnSaksB", "LogginnSaksB")">Logg inn SaksB</a>
-                            <a href="@Url.Action("SortereInnmeldingerSaksB", "SortereInnmeldingerSaksB")">Sortere Innmeldinger SaksB</a>
-                            <a href="@Url.Action("SøkEtterInnmelderSaksB", "SøkEtterInnmelderSaksB")">Søke etter innmelde SaksBr</a>
-                            <a href="@Url.Action("KartvisningAlleInnmeldingerSaksB", "KartvisningAlleInnmeldingerSaksB")">KartvisningAlleInnmeldingerSaksBr</a>
+                            <a href="@Url.Action("SortereInnmeldingerSaksB", "SortereInnmeldingerSaksB")">SortereInnmeldingerSaksB</a>
+                            <a href="@Url.Action("SøkEtterInnmelderSaksB", "SøkEtterInnmelderSaksB")">SøkeEtterInnmelderSaksB</a>
+                            <a href="@Url.Action("KartvisningAlleInnmeldingerSaksB", "KartvisningAlleInnmeldingerSaksB")">KartvisningAlleInnmeldingerSaksB</a>
+                            <a href="@Url.Action("DropDownDummySaksB", "DropDownDummySaksB")">DropDownDummySaksB</a>
                             
                         </div>
                     </ul>
