+++ ViewModels/BehandleInnmeldingSaksBViewModel.cs
@@ -6,8 +6,7 @@ namespace ViewModels
 {
     public class BehandleInnmeldingSaksBViewModel
     {
-        public InnmeldingEModel InnmeldingE { get; set; }
-        public List<SelectListItem> StatusList { get; set; }
+       
         
     }
 }
